// frontend/src/modules/Auth/Auth.enums.ts
export enum FormHeaderLabel {
  Submit = 'Submit',
  ResetPassword = 'Reset Password',
  SignIn = 'Sign-In',
  SignUp = 'Sign-Up',
  SignInViaGoogle = 'Sign-In via Google',
  SignUpViaGoogle = 'Sign-Up via Google',
  ForgotPassword = 'Forgot Password',
}

export enum FormLinkLabel {
  SignUpLink = "Don't have an account?",
  LoginLink = 'Already have an account?',
  ForgetPasswordLink = 'Forgot a password?',
}